+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Release 1.0.2
 
+#### New Components
+ - InlineInput for inlining Items.
+ - InlineFileInput for inline FileItems.
+ - InlineGenericInput for inlining Items using keys.
+
 #### Features
  - Update GithubRepo to use a dataclass for metadata to improve type safety/checking.
- - Updated debug logging to be more readable and include appropriate information
+ - Updated debug logging to be more readable and include appropriate information.
 
 ## Release 1.0.1